Trying to split components into state, logic and view
View should be the part instantiated in the main thread. Contains rendering information.
Logic should be instantiated in the web worker. Contains all the business logic.
State contains the properties of the object. It's syncrhonized between the main thread and the web worker.

It failed because synchronization is too complex and even if it's solved each component needs it's View or Logic representation for the state so it's another layer for syncrhonization